refactor(engine): Condense dsl module
