+++ tracecat/experimental/dsl/trigger.py
@@ -1,6 +1,5 @@
 import asyncio
 import json
-import logging
 import sys
 import uuid
 
@@ -9,6 +8,7 @@ from temporalio.client import Client
 
 from tracecat.experimental.dsl._converter import pydantic_data_converter
 from tracecat.experimental.dsl.workflow import DSLInput, DSLWorkflow
+from tracecat.logging import logger
 
 
 async def main(dsl_yaml: str) -> None:
@@ -29,12 +29,10 @@ async def main(dsl_yaml: str) -> None:
         id=f"dsl-workflow-id-{uuid.uuid4()}",
         task_queue="dsl-task-queue",
     )
-    logging.info(f"Workflow result:\n{json.dumps(result, indent=2)}")
+    logger.info(f"Workflow result:\n{json.dumps(result, indent=2)}")
 
 
 if __name__ == "__main__":
-    logging.basicConfig(level=logging.INFO)
-
     # Require the YAML file as an argument. We read this _outside_ of the async
     # def function because thread-blocking IO should never happen in async def
     # functions.
