+++ tracecat/experimental/dsl/workflow.py
@@ -4,15 +4,14 @@ import asyncio
 from datetime import timedelta
 from typing import Any, TypedDict
 
-from temporalio import workflow
-
-from tracecat.experimental.dsl.activities import DSLActivities
+from temporalio import activity, workflow
 
 with workflow.unsafe.imports_passed_through():
     import jsonpath_ng.lexer  # noqa
     import jsonpath_ng.parser  # noqa
     from pydantic import BaseModel, ConfigDict, Field
 
+    from tracecat.experimental.actions import registry
     from tracecat.experimental.templates.eval import eval_templated_object
 
 
@@ -26,12 +25,12 @@ class DSLInput(_DSLNode):
 
 
 SLUG_PATTERN = r"^[a-z0-9_]+$"
+ACTION_TYPE_PATTERN = r"^[a-z0-9_.]+$"
 
 
 class ActivityStatement(_DSLNode):
     ref: str = Field(pattern=SLUG_PATTERN)
-    action: str = Field(pattern=SLUG_PATTERN)
-    # Lower snake case slug is used for the key
+    action: str = Field(pattern=ACTION_TYPE_PATTERN)
     args: dict[str, Any] = Field(default_factory=dict)
 
 
@@ -70,17 +69,15 @@ class DSLWorkflow:
 
     async def execute_statement(self, stmt: Statement) -> None:
         match stmt:
+            # Action is the qualname of the udf
             case ActivityStatement(ref=ref, action=action, args=args):
                 # Resolve all templated arguments
                 processed_args = eval_templated_object(args, operand=self.context)
 
-                # Validate the arguments
-                validator = DSLActivities.validators[action]
-                # We use strict=True to ensure that no type coercion is done
-                validated_args = validator.model_validate(processed_args, strict=True)
+                # TODO: Set a retry policy for the activity
                 activity_result = await workflow.execute_activity(
-                    action,
-                    arg=validated_args,
+                    "run_udf",
+                    arg=UDFAction(type=action, args=processed_args),
                     start_to_close_timeout=timedelta(minutes=1),
                 )
                 self.context["ACTIONS"][ref] = DSLNodeResult(
@@ -99,3 +96,35 @@ class DSLWorkflow:
                         tg.create_task(self.execute_statement(branch))
             case _:
                 raise NotImplementedError(f"Unsupported statement: {stmt}")
+
+
+class UDFAction(BaseModel):
+    type: str
+    args: dict[str, Any]
+
+
+class DSLActivities:
+    def __new__(cls):
+        raise RuntimeError("This class should not be instantiated")
+
+    @staticmethod
+    @activity.defn
+    async def run_udf(input: UDFAction) -> Any:
+        activity.logger.info("Run udf")
+        activity.logger.info(f"{input = }")
+        udf = registry[input.type]
+        if udf.is_async:
+            result = await udf.fn(**input.args)
+        else:
+            # Force it to be async
+            result = await asyncio.to_thread(udf.fn, **input.args)
+        activity.logger.info(f"Result: {result}")
+        return result
+
+
+# Dynamically register all static methods as activities
+dsl_activities = [
+    getattr(DSLActivities, method_name)
+    for method_name in dir(DSLActivities)
+    if hasattr(getattr(DSLActivities, method_name), "__temporal_activity_definition")
+]
